requirements.txt
Cython==0.19.1
Django==1.5.1
Jinja2==2.7
MarkupSafe==0.18
Pillow==2.1.0
Pygments==1.6
South==0.7.6
Sphinx==1.2b1
argparse==1.2.1
bpython==0.12
coverage==3.6
distribute==0.6.48
dj-database-url==0.2.1
dj-static==0.0.5
django-activity-stream==0.4.5beta1
django-braces==1.0.0
django-debug-toolbar==0.9.4
django-discover-runner==0.4
django-jsonfield==0.9.10
django-model-utils==1.3.1
django-registration==1.0
django-social-auth==0.7.23
django-timezones==0.2
django-toolbelt==0.0.1
docutils==0.10
dota2py==0.1.2
easy-thumbnails==1.3
-e git://github.com/surfly/gevent.git@454a77ca561868854760b2d9cbfa3bf3bbd2e062#egg=gevent-dev
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.1
gunicorn==17.5
httplib2==0.8
logutils==0.3.3
nose==1.3.0
oauth2==1.5.211
psycopg2==2.5.1
python-openid==2.2.5
pytz==2013b
redis==2.7.6
requests==1.2.3
simplejson==3.3.0
static==0.4
wsgiref==0.1.2
